pyproject.toml
[project]
name = "bampi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bampy",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot2[fastapi]>=2.4.4",
    "openai>=1.0",
]

[tool.nonebot]
plugins = []
plugin_dirs = ["bampi/plugins"]

[tool.uv.workspace]
members = [
    "bampy",
]

[tool.uv.sources]
bampy = { workspace = true, editable = true }
